# New evaluator that doesn't depend on flattening