--- a/quint/src/cli.ts
+++ b/quint/src/cli.ts
@@ -256,7 +256,7 @@ const runCmd = {
       .option('invariant', {
         desc: 'invariant to check: a definition name or an expression',
         type: 'string',
-        default: ['true'],
+        default: 'true',
       })
       .option('seed', {
         desc: 'random seed to use for non-deterministic choice',