--- a/quint/src/cliCommands.ts
+++ b/quint/src/cliCommands.ts
@@ -17,6 +17,7 @@ import {
   SourceMap,
   compactSourceMap,
   parseDefOrThrow,
+  parseExpressionOrDeclaration,
   parsePhase1fromText,
   parsePhase2sourceResolution,
   parsePhase3importAndNameResolution,
@@ -24,19 +25,29 @@ import {
 } from './parsing/quintParserFrontend'
 import { ErrorMessage } from './ErrorMessage'
 
-import { Either, left, right } from '@sweet-monads/either'
+import { Either, left, mergeInMany, right } from '@sweet-monads/either'
 import { fail } from 'assert'
 import { EffectScheme } from './effects/base'
-import { LookupTable, UnusedDefinitions } from './names/base'
+import { LookupTable, NameResolutionResult, UnusedDefinitions } from './names/base'
 import { ReplOptions, quintRepl } from './repl'
-import { FlatModule, OpQualifier, QuintEx, QuintModule, QuintOpDef, qualifier } from './ir/quintIr'
+import {
+  FlatModule,
+  OpQualifier,
+  QuintApp,
+  QuintBool,
+  QuintEx,
+  QuintModule,
+  QuintOpDef,
+  isDef,
+  qualifier,
+} from './ir/quintIr'
 import { TypeScheme } from './types/base'
 import { createFinders, formatError } from './errorReporter'
 import { DocumentationEntry, produceDocs, toMarkdown } from './docs'
 import { QuintError, quintErrorToString } from './quintError'
-import { TestOptions, TestResult, compileAndTest } from './runtime/testing'
+import { TestOptions, TestResult } from './runtime/testing'
 import { IdGenerator, newIdGenerator } from './idGenerator'
-import { SimulatorOptions, compileAndRun } from './simulation'
+import { Outcome, SimulatorOptions } from './simulation'
 import { ofItf, toItf } from './itf'
 import { printExecutionFrameRec, printTrace, terminalWidth } from './graphics'
 import { verbosity } from './verbosity'
@@ -45,10 +56,13 @@ import { fileSourceResolver } from './parsing/sourceResolver'
 import { verify } from './quintVerifier'
 import { flattenModules } from './flattening/fullFlattener'
 import { AnalysisOutput, analyzeInc, analyzeModules } from './quintAnalyzer'
-import { ExecutionFrame } from './runtime/trace'
+import { ExecutionFrame, newTraceRecorder } from './runtime/trace'
 import { flow, isEqual, uniqWith } from 'lodash'
 import { Maybe, just, none } from '@sweet-monads/maybe'
 import { compileToTlaplus } from './compileToTlaplus'
+import { Evaluator } from './runtime/impl/evaluator'
+import { NameResolver } from './names/resolver'
+import { walkExpression } from './ir/IRVisitor'
 
 export type stage =
   | 'loading'
@@ -65,7 +79,7 @@ interface OutputStage {
   stage: stage
   // the modules and the lookup table produced by 'parse'
   modules?: QuintModule[]
-  table?: LookupTable
+  names?: NameResolutionResult
   unusedDefinitions?: UnusedDefinitions
   // the tables produced by 'typecheck'
   types?: Map<bigint, TypeScheme>
@@ -91,7 +105,7 @@ const pickOutputStage = ({
   stage,
   warnings,
   modules,
-  table,
+  names,
   unusedDefinitions,
   types,
   effects,
@@ -107,7 +121,7 @@ const pickOutputStage = ({
     stage,
     warnings,
     modules,
-    table,
+    names,
     unusedDefinitions,
     types,
     effects,
@@ -137,6 +151,7 @@ interface ParsedStage extends LoadedStage {
   sourceMap: SourceMap
   table: LookupTable
   unusedDefinitions: UnusedDefinitions
+  resolver: NameResolver
   idGen: IdGenerator
 }
 
@@ -321,6 +336,7 @@ export async function runTests(prev: TypecheckedStage): Promise<CLIProcedure<Tes
   const rng = rngOrError.unwrap()
 
   const matchFun = (n: string): boolean => isMatchingTest(testing.args.match, n)
+  const maxSamples = testing.args.maxSamples
   const options: TestOptions = {
     testMatch: matchFun,
     maxSamples: testing.args.maxSamples,
@@ -351,69 +367,24 @@ export async function runTests(prev: TypecheckedStage): Promise<CLIProcedure<Tes
     out(`\n  ${mainName}`)
   }
 
-  // TODO: This block is a workaround for the fact that flattening removes any defs
-  // not refernced in the main module. We'd instead like way to just instruct it
-  // to keep some defs.
-  //
-  // Find tests that are not used in the main module. We need to add references to them in the main module so flattening
-  // doesn't drop the definitions.
-  const unusedTests = [...testing.unusedDefinitions(mainName)].filter(
-    d => d.kind === 'def' && options.testMatch(d.name)
-  )
-  // Define name expressions referencing each test that is not referenced elsewhere, adding the references to the lookup
-  // table
-  const args: QuintEx[] = unusedTests.map(defToAdd => {
-    const id = testing.idGen.nextId()
-    testing.table.set(id, defToAdd)
-    const namespace = defToAdd.importedFrom ? qualifier(defToAdd.importedFrom) : undefined
-    const name = namespace ? [namespace, defToAdd.name].join('::') : defToAdd.name
+  const main = prev.modules.find(m => m.name === mainName)!
+  const testDefs = main.declarations.filter(d => d.kind === 'def' && options.testMatch(d.name)) as QuintOpDef[]
 
-    return { kind: 'name', id, name }
+  const evaluator = new Evaluator(testing.table, newTraceRecorder(verbosityLevel, rng, 1), rng)
+  const results = testDefs.map(def => {
+    return evaluator.test(def.name, def.expr, maxSamples)
   })
-  // Wrap all expressions in a single declaration
-  const testDeclaration: QuintOpDef = {
-    id: testing.idGen.nextId(),
-    name: 'q::unitTests',
-    kind: 'def',
-    qualifier: 'run',
-    expr: { kind: 'app', opcode: 'actionAll', args, id: testing.idGen.nextId() },
-  }
-  // Add the declaration to the main module
-  const main = testing.modules.find(m => m.name === mainName)
-  main?.declarations.push(testDeclaration)
-
-  // TODO The following block should all be moved into compileAndTest
-  const analysisOutput = { types: testing.types, effects: testing.effects, modes: testing.modes }
-  const { flattenedModules, flattenedTable, flattenedAnalysis } = flattenModules(
-    testing.modules,
-    testing.table,
-    testing.idGen,
-    testing.sourceMap,
-    analysisOutput
-  )
-  const compilationState = {
-    originalModules: testing.modules,
-    modules: flattenedModules,
-    sourceMap: testing.sourceMap,
-    analysisOutput: flattenedAnalysis,
-    idGen: testing.idGen,
-    sourceCode: testing.sourceCode,
-  }
-
-  const testResult = compileAndTest(compilationState, mainName, flattenedTable, options)
 
   // We have a compilation failure, so early exit without reporting test results
-  if (testResult.isLeft()) {
-    return cliErr('Tests could not be run due to an error during compilation', {
-      ...testing,
-      errors: testResult.value.map(mkErrorMessage(testing.sourceMap)),
-    })
-  }
+  // if (testResult.isLeft()) {
+  //   return cliErr('Tests could not be run due to an error during compilation', {
+  //     ...testing,
+  //     errors: testResult.value.map(mkErrorMessage(testing.sourceMap)),
+  //   })
+  // }
 
   // We're finished running the tests
   const elapsedMs = Date.now() - startMs
-  // So we can analyze the results now
-  const results = testResult.value
 
   // output the status for every test
   let nFailures = 1
@@ -529,6 +500,7 @@ function maybePrintCounterExample(verbosityLevel: number, states: QuintEx[], fra
  */
 export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure<TracingStage>> {
   const simulator = { ...prev, stage: 'running' as stage }
+  const startMs = Date.now()
   const verbosityLevel = deriveVerbosity(prev.args)
   const mainName = guessMainModule(prev)
 
@@ -563,35 +535,66 @@ export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure
     },
   }
 
-  const startMs = Date.now()
+  const recorder = newTraceRecorder(options.verbosity, options.rng, options.numberOfTraces)
 
-  const mainText = prev.sourceCode.get(prev.path)!
-  const mainPath = fileSourceResolver(prev.sourceCode).lookupPath(dirname(prev.args.input), basename(prev.args.input))
-  const mainModule = prev.modules.find(m => m.name === mainName)
-  if (mainModule === undefined) {
-    return cliErr(`Main module ${mainName} not found`, { ...simulator, errors: [] })
+  function toExpr(input: string): Either<QuintError, QuintEx> {
+    const parseResult = parseExpressionOrDeclaration(input, '<input>', prev.idGen, prev.sourceMap)
+    if (parseResult.kind !== 'expr') {
+      return left({ code: 'QNT501', message: `Expected ${input} to be a valid expression` })
+    }
+
+    prev.resolver.switchToModule(mainName)
+    walkExpression(prev.resolver, parseResult.expr)
+    if (prev.resolver.errors.length > 0) {
+      return left(prev.resolver.errors[0])
+    }
+
+    return right(parseResult.expr)
   }
-  const mainId = mainModule.id
-  const mainStart = prev.sourceMap.get(mainId)!.start.index
-  const mainEnd = prev.sourceMap.get(mainId)!.end!.index
-  const result = compileAndRun(newIdGenerator(), mainText, mainStart, mainEnd, mainName, mainPath, options)
+
+  const argsParsingResult = mergeInMany([prev.args.init, prev.args.step, prev.args.invariant].map(toExpr))
+  if (argsParsingResult.isLeft()) {
+    return cliErr('Argument error', {
+      ...simulator,
+      errors: argsParsingResult.value.map(mkErrorMessage(new Map())),
+    })
+  }
+  const [init, step, invariant] = argsParsingResult.value
+
+  const evaluator = new Evaluator(prev.resolver.table, recorder, options.rng)
+  const evalResult = evaluator.simulate(
+    init,
+    step,
+    invariant,
+    prev.args.maxSamples,
+    prev.args.maxSteps,
+    prev.args.nTraces ?? 1,
+    prev.effects
+  )
 
   const elapsedMs = Date.now() - startMs
 
-  switch (result.outcome.status) {
+  const outcome: Outcome = evalResult.isRight()
+    ? { status: (evalResult.value as QuintBool).value ? 'ok' : 'violation' }
+    : { status: 'error', errors: [evalResult.value] }
+
+  const states = recorder.bestTraces[0].frame.args
+  const frames = recorder.bestTraces[0].frame.subframes
+  const seed = options.rng.getState()
+  switch (outcome.status) {
     case 'error':
       return cliErr('Runtime error', {
         ...simulator,
-        status: result.outcome.status,
-        trace: result.states,
-        errors: result.outcome.errors.map(mkErrorMessage(prev.sourceMap)),
+        status: outcome.status,
+        trace: states,
+        errors: outcome.errors.map(mkErrorMessage(prev.sourceMap)),
       })
 
     case 'ok':
-      maybePrintCounterExample(verbosityLevel, result.states, result.frames)
+      maybePrintCounterExample(verbosityLevel, states, frames)
       if (verbosity.hasResults(verbosityLevel)) {
         console.log(chalk.green('[ok]') + ' No violation found ' + chalk.gray(`(${elapsedMs}ms).`))
-        console.log(chalk.gray(`Use --seed=0x${result.seed.toString(16)} to reproduce.`))
+        console.log(chalk.gray(`Use --seed=0x${seed.toString(16)} to reproduce.`))
         if (verbosity.hasHints(verbosityLevel)) {
           console.log(chalk.gray('You may increase --max-samples and --max-steps.'))
           console.log(chalk.gray('Use --verbosity to produce more (or less) output.'))
@@ -600,15 +603,15 @@ export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure
 
       return right({
         ...simulator,
-        status: result.outcome.status,
-        trace: result.states,
+        status: outcome.status,
+        trace: states,
       })
 
     case 'violation':
-      maybePrintCounterExample(verbosityLevel, result.states, result.frames)
+      maybePrintCounterExample(verbosityLevel, states, frames)
       if (verbosity.hasResults(verbosityLevel)) {
         console.log(chalk.red(`[violation]`) + ' Found an issue ' + chalk.gray(`(${elapsedMs}ms).`))
-        console.log(chalk.gray(`Use --seed=0x${result.seed.toString(16)} to reproduce.`))
+        console.log(chalk.gray(`Use --seed=0x${seed.toString(16)} to reproduce.`))
 
         if (verbosity.hasHints(verbosityLevel)) {
           console.log(chalk.gray('Use --verbosity=3 to show executions.'))
@@ -617,8 +620,8 @@ export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure
 
       return cliErr('Invariant violated', {
         ...simulator,
-        status: result.outcome.status,
-        trace: result.states,
+        status: outcome.status,
+        trace: states,
         errors: [],
       })
   }