--- a/quint/src/names/base.ts
+++ b/quint/src/names/base.ts
@@ -252,4 +252,6 @@ export const builtinNames = [
   'variant',
   'q::debug',
   'q::lastTrace',
+  'q::test',
+  'q::testOnce',
 ]