… CHANGELOG.md
… quint/io-cli-tests.md
… quint/src/names/resolver.ts
… quint/src/parsing/quintParserFrontend.ts
… quint/src/repl.ts
… quint/src/runtime/compile.ts
… quint/src/runtime/impl/Context.ts
… quint/src/runtime/impl/VarStorage.ts
… quint/src/runtime/impl/base.ts
… quint/src/runtime/impl/builder.ts
… quint/src/runtime/impl/builtins.ts
… quint/src/runtime/impl/compiler.ts
… quint/src/runtime/impl/compilerImpl.ts
… quint/src/runtime/impl/evaluator.ts
… quint/src/runtime/impl/operatorEvaluator.ts
… quint/src/runtime/impl/runtimeValue.ts
… quint/src/runtime/impl/trace.ts
… quint/src/runtime/runtime.ts
… quint/src/runtime/testing.ts
… quint/src/runtime/trace.ts
… quint/src/simulation.ts
… quint/src/static/callgraph.ts
… quint/src/types/builtinSignatures.ts
… quint/test/flattening/flattener.test.ts
… quint/test/repl.test.ts
… quint/test/runtime/compile.test.ts
… quint/test/runtime/trace.test.ts
… quint/test/static/callgraph.test.ts
… quint/testFixture/SuperSpec.json
… quint/testFixture/_1016nonConstOverride.json